requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
diskcache==5.6.3
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llama_cpp_python==0.3.16
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
